Conflicts:
	app/conf/navigation.conf
	app/helpers/utilityHelpers.php
	app/lib/ca/BaseLookupController.php
	app/lib/ca/Browse/BrowseEngine.php
	app/lib/ca/ConfigurationCheck.php
	app/lib/core/BaseModel.php
	app/lib/core/Db.php
	app/lib/core/Db/DbStatement.php
	app/lib/core/Db/pgsql.php
	app/lib/core/Plugins/SearchEngine/DidYouMean.php
	app/models/ca_commerce_orders.php
	app/models/ca_entities.php
	app/models/ca_list_items.php
	app/models/ca_lists.php
	app/models/ca_object_representations.php
	app/models/ca_places.php
	install/inc/Installer.php
	install/inc/SchemaLoader.php
	install/inc/schema_mysql.sql
	setup.php-dist
	themes/default/css/base.css

+++ install/inc/Installer.php
@@ -282,7 +282,7 @@ class Installer {
 		$vo_dm = Datamodel::load();
 		$vo_db = new Db();
 		$vs_dbtype = $vo_config->get("db_type");
-		/*if (defined('__CA_ALLOW_INSTALLER_TO_OVERWRITE_EXISTING_INSTALLS__') && __CA_ALLOW_INSTALLER_TO_OVERWRITE_EXISTING_INSTALLS__ && ($this->opb_overwrite)) {
+		if (defined('__CA_ALLOW_INSTALLER_TO_OVERWRITE_EXISTING_INSTALLS__') && __CA_ALLOW_INSTALLER_TO_OVERWRITE_EXISTING_INSTALLS__ && ($this->opb_overwrite) && ($vs_dbtype != "pgsql")) {
 			$vo_db->query('DROP DATABASE IF EXISTS '.__CA_DB_DATABASE__);
 			$vo_db->query('CREATE DATABASE '.__CA_DB_DATABASE__);
 			$vo_db->query('USE '.__CA_DB_DATABASE__);
